refactor: GrammarNumOfGrammarBookDto의 grammarNum 타입 변경

+++ src/main/java/com/wordwave/grammarbook/dto/GrammarNumOfGrammarBookDto.java
@@ -14,5 +14,5 @@ public class GrammarNumOfGrammarBookDto {
 
     private String name;
 
-    private Integer grammarNum;
+    private Long grammarNum;
 }
